move report hooks utils
