+++ apps/fishing-map/features/map/Map.tsx
@@ -29,7 +29,7 @@ import { useMapLayers } from 'features/map/map-layers.hooks'
 import MapPopups from 'features/map/popups/MapPopups'
 import { MapCoordinates } from 'types'
 import { useAppDispatch } from 'features/app/app.hooks'
-import { useHasReportTimeseries } from 'features/reports/areas/reports-timeseries.hooks'
+import { useHasReportTimeseries } from 'features/reports/activity/reports-activity-timeseries.hooks'
 import { selectReportAreaStatus } from 'features/reports/areas/reports.selectors'
 import { AsyncReducerStatus } from 'utils/async-slice'
 import {
