+++ apps/fishing-map/features/reports/activity/ReportActivityEvolution.tsx
@@ -16,7 +16,7 @@ import { FourwingsInterval } from '@globalfishingwatch/deck-loaders'
 import i18n from 'features/i18n/i18n'
 import { formatDateForInterval, getUTCDateTime } from 'utils/dates'
 import { formatI18nNumber } from 'features/i18n/i18nNumber'
-import { formatEvolutionData } from 'features/reports/areas/reports-timeseries.utils'
+import { formatEvolutionData } from 'features/reports/activity/reports-activity-timeseries.utils'
 import { ReportActivityProps } from 'features/reports/activity/ReportActivityGraph'
 import { tickFormatter } from '../areas/reports.utils'
 import styles from './ReportActivityEvolution.module.css'
