+++ apps/fishing-map/features/reports/activity/ReportActivityGraphSelector.tsx
@@ -10,7 +10,7 @@ import {
 } from 'data/config'
 import { selectActiveReportDataviews } from 'features/dataviews/selectors/dataviews.selectors'
 import { useFitAreaInViewport } from 'features/reports/areas/reports.hooks'
-import { useSetReportTimeComparison } from 'features/reports/areas/reports-timecomparison.hooks'
+import { useSetReportTimeComparison } from 'features/reports/activity/reports-activity-timecomparison.hooks'
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
 import { selectReportActivityGraph } from '../areas/reports.config.selectors'
 import { ReportActivityGraph } from '../areas/reports.types'
