+++ apps/fishing-map/features/reports/activity/ReportActivityPeriodComparison.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
 import { InputDate, InputText, Select } from '@globalfishingwatch/ui-components'
-import { useReportTimeCompareConnect } from 'features/reports/areas/reports-timecomparison.hooks'
+import { useReportTimeCompareConnect } from 'features/reports/activity/reports-activity-timecomparison.hooks'
 import { selectActiveActivityAndDetectionsDataviews } from 'features/dataviews/selectors/dataviews.selectors'
 import { getSourcesSelectedInDataview } from 'features/workspace/activity/activity.utils'
 import { selectReportAreaIds } from 'features/reports/areas/reports.selectors'
