+++ apps/fishing-map/features/reports/activity/reports-activity-geo.utils.workers.ts
@@ -1,4 +1,4 @@
-import { filterByPolygon, FilterByPolygomParams } from './reports-geo.utils'
+import { filterByPolygon, FilterByPolygomParams } from './reports-activity-geo.utils'
 
 // eslint-disable-next-line no-restricted-globals
 addEventListener('message', (event: MessageEvent<FilterByPolygomParams>) => {
