+++ apps/fishing-map/features/reports/areas/title/ReportTitle.tsx
@@ -38,7 +38,7 @@ import { BufferOperation, BufferUnit } from 'types'
 import { cleanCurrentWorkspaceStateBufferParams } from 'features/workspace/workspace.slice'
 import { AsyncReducerStatus } from 'utils/async-slice'
 import { formatI18nNumber } from 'features/i18n/i18nNumber'
-import { useReportFeaturesLoading } from 'features/reports/areas/reports-timeseries.hooks'
+import { useReportFeaturesLoading } from 'features/reports/activity/reports-activity-timeseries.hooks'
 import { useHighlightReportArea } from '../reports.hooks'
 import { BufferButtonTooltip } from './BufferButonTooltip'
 import styles from './ReportTitle.module.css'
